Change Sprite Unique ID to a long int.
A long running game could reach up to 4294967296 sprites.
But 18446744073709551616 sprites should be enough for anyone.

+++ Source/Sprites/sprite.cpp
@@ -17,7 +17,7 @@
  */
 
 // Sprite ID 0 is only used as a NULL
-int Sprite::sprite_ids = 1;
+long int Sprite::sprite_ids = 1;
 
 /**\class Sprite
  * \brief Supertype for all objects in the world.
